# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NEXUS — API del Sistema Principal
  description: |
    API del módulo `SP`. **Toda operación exige un token de acceso**
    salvo el inicio, la renovación y el cierre de sesión.

    Para probar cualquier endpoint: obtenga el token con
    `POST /api/v1/auth/login`, pulse **Authorize** y pegue el valor de
    `accessToken` — sin el prefijo `Bearer`, que Swagger añade solo.

    El token vive **quince minutos** y no es revocable: cuando expire,
    renuévelo con `POST /api/v1/auth/refresh` y vuelva a autorizar.
  version: v1
servers:
- url: /
  description: Este mismo servidor
security:
- bearerAuth: []
...
paths:
...
  /api/v1/user-commission-rates/{id}/products/{productId}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar la asociación de una tasa personalizada con un producto
      description: |
        **Deja de regir ahí, y la tasa sigue viva.** Puede volver a asociarse a
        ese producto o a otro; retirarla es otra operación.

        **El motivo es obligatorio.** La fila se borra de verdad —una asociación
        es configuración vigente, no un hecho del pasado— de modo que este texto
        es **el único sitio** donde quedará escrito por qué esa persona dejó de
        tener su excepción en ese producto.

        **`404` y no `409` si ya no estaba**: con el borrado físico no queda nada
        que distinga «nunca existió» de «ya se borró», y un `409` afirmaría algo
        que no se sabe.
      operationId: desasociar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissociateProductRequest"
      responses:
        "204":
          description: Desasociada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: Esa tasa no está asociada a ese producto
  /api/v1/user-commission-rates/{id}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar una tasa personalizada
      description: |
        Retira una tasa que **no debió existir**, con **motivo obligatorio**.

        **Retirar no es cerrar la vigencia**: se cierra lo que dejó de regir, se
        retira lo que fue un error. Por eso el retiro **no toca la vigencia** — el
        registro de eliminación debe poder decir qué periodo cubría lo retirado.

        Los días que ocupaba **quedan libres**, de modo que puede declararse otra
        tasa que los cubra.

        **No es idempotente**: retirar dos veces devuelve `409`.
      operationId: retirar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCommissionRateRequest"
      responses:
        "204":
          description: Tasa retirada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: La tasa no existe
        "409":
          description: Ya estaba retirada
...
  /api/v1/products:
    get:
      tags:
      - Productos
      summary: Consultar el catálogo de productos
      description: |
        Devuelve el catálogo **paginado**, con el destino de cada upgrade y la
        moneda de cada precio ya resueltos: no hace falta una segunda consulta.

        **Los retirados quedan fuera salvo que se pidan** con `includeDeleted`.
        Al pedirlos se indica **desde cuándo** lo están; el **motivo** no viaja
        en el listado —uno a uno lo devuelve el detalle (`RF-PM-003`), en
        bloque sería una exportación de decisiones comerciales—. Verlos **no
        exige un permiso propio**: basta `products:read`.

        Cada fila trae **los dos precios**: `price` —el que se cobra— y
        `purchasePrice` —el **precio de compra**, lo que NEXUS paga por el
        producto—, este **presente y nulo** en los productos cuyo costo no se
        conoce. Este listado y el detalle son **los dos únicos sitios** donde
        se ven juntos: la oferta y el hotlink no devuelven el de compra
        (`RN-PM-024`, 12-09-2026).

        Y trae **`exchange`**, la conversión a la moneda por omisión con la
        tasa vigente hoy, calculada **sobre `price`** —el de compra nunca se
        convierte—. Llega **presente y nula** cuando el producto ya está en
        esa moneda o cuando nadie declaró una tasa: eso **no es un error** y
        el producto se devuelve igual.

        Cada fila trae también **`videoUrl`**, la dirección del video que
        presenta el producto, **tal cual se guardó** y **presente y nula**
        cuando no tiene (`RN-PM-032`). **No es un filtro.**

        Solo se puede ordenar por la lista blanca —`name`, `price`,
        `createdAt`—, con `,asc` o `,desc`. **`purchasePrice` no está en
        ella**: ordenar por una columna que admite nulos obligaría a decidir
        dónde van los productos sin costo conocido, y nadie lo ha decidido.
        **Un campo fuera de la lista se rechaza y no se ignora**:
        ignorarlo devolvería un orden distinto del pedido sin decirlo. Por omisión se ordena por **fecha de alta
        descendente**, y el orden aplicado viaja en la respuesta.

        **`targetMembershipId` no se valida contra el catálogo de membresías.**
        Filtrar por un destino inexistente devuelve la colección vacía y no es
        un error; combinarlo con `type=BOT` también, porque ningún
        bot tiene destino.

        La búsqueda va sobre el nombre, **sin distinguir acentos ni
        mayúsculas** y por fragmento. En blanco equivale a no filtrar.

        **`scope` e `implementation` filtran como `type` y `status`**: se
        admiten en cualquier caja y un valor fuera de dominio se rechaza junto
        al resto de parámetros inválidos, no en una vuelta aparte. El de
        alcance es **el único sitio donde ese dato se consulta hoy** — la
        oferta de `RF-PM-007` no filtra por él.

        Un filtro sin coincidencias devuelve `200` con la colección vacía, y
        una página más allá de la última hace lo mismo **con el total real**.
        No hay `404` ni `422`.
      operationId: listar_3
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: scope
        in: query
        required: false
        schema:
          type: string
      - name: implementation
        in: query
        required: false
        schema:
          type: string
      - name: sourceMembershipId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: targetMembershipId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: "Página del catálogo, con el orden aplicado."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductPageResponse"
        "400":
          description: "Paginación fuera de rango (`VAL-001`), tipo (`VAL-002`) o\
            \ estado (`VAL-003`) fuera de su dominio, identificador mal formado (`VAL-004`)\
            \ o campo de ordenamiento no admitido (`VAL-005`). Los cuatro primeros\
            \ se devuelven **juntos**"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin el permiso `products:read` (`AUTH-002`)
        "500":
          description: Fallo no controlado (`ERR-500`)
    post:
      tags:
      - Productos
      summary: Registrar un producto
      description: |
        Registra un producto del catálogo, **siempre inactivo**: publicarlo es otra
        operación (`RF-PM-005`).

        El tipo decide qué campos son obligatorios: un `UPGRADE_MEMBRESIA` debe
        declarar **las dos** membresías —de qué nivel sale y a cuál lleva— y un
        `BOT` no puede declarar ninguna.

        **El origen no puede estar por encima del destino**: eso sería vender un
        descenso llamándolo upgrade, y se rechaza con `422` (`EX-006`). **Sí
        puede ser el MISMO**, y entonces el producto es una **renovación**: lo
        que vende es tiempo —su `validityDays`— y no un cambio de nivel.

        **Y no tiene por qué ser el inmediatamente inferior**: saltar niveles es
        legítimo, y es la razón de que el origen se declare en lugar de
        deducirse de la cadena.

        `icon` es el **nombre** del icono con el que el frontend pinta el producto
        —no una imagen—, en minúsculas y guion medio. Es **opcional**, y solo un
        `UPGRADE_MEMBRESIA` puede llevarlo: en un `BOT` se rechaza (`RN-PM-016`).

        `videoUrl` es **la dirección de un video** que presenta el producto —no
        el video—. Es **opcional y vale en los dos tipos**, sin la condición
        del icono. Se comprueba **solo la forma**: una URL absoluta `http` o
        `https`, sin espacios y de hasta 500 caracteres; lo demás se rechaza
        con `VAL-017`. **El sistema no sigue el enlace** —no comprueba que el
        video exista ni lo descarga—, y lo guarda **tal cual se escribió**,
        recortado y sin normalizar nada más. Ausente o nulo significan lo
        mismo: no tiene video. **Al revés que el precio de compra, sale en las
        cuatro lecturas**, el hotlink sin token incluido (`RN-PM-032`).

        La vigencia es opcional en los dos tipos: sin ella, lo adquirido no caduca.

        `scope` e `implementation` son **obligatorios y en los dos tipos**, y
        **no tienen valor por omisión**: `scope` dice hasta dónde se muestra el
        producto —`HOTLINKS` **incluye** `TIENDA`, no la sustituye— e
        `implementation` dice si lo comprado se aplica solo (`AUTOMATICA`) o
        espera a que un funcionario lo autorice (`MANUAL`).

        **Un producto lleva DOS precios y solo uno se cobra.** `price` es el
        del sistema —el que copia la venta y sobre el que se comisiona— y
        `purchasePrice` es el **precio de compra**: lo que NEXUS paga por el
        producto cuando tiene que comprarlo, y donde se guarda lo que costó.
        Es **opcional**, no interviene en ningún cálculo, se expresa en la
        **misma moneda** y **no sale de administración**: la oferta y el
        hotlink no lo devuelven. Ausente o nulo significan lo mismo —no se
        conoce todavía—, y **eso no es «costó cero»** (`RN-PM-023`). Se llamó
        `publicPrice` hasta el 12-09-2026; ese nombre es hoy una propiedad
        desconocida y devuelve `400`.

        Los dos importes se validan igual: **no negativos** —el **cero se
        admite** desde que existe la renovación de una membresía gratuita— y
        con los decimales que declare su moneda. El rechazo **nombra el
        campo** que incumple.

        **Los dos solo se ven desde administración** (`RN-PM-024`): la oferta
        de un cliente y el hotlink público devuelven **uno**.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterProductRequest"
        required: true
      responses:
        "201":
          description: "Producto registrado, en estado `INACTIVO`."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Datos inválidos (serie `VAL-nnn`)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "403":
          description: Autenticado sin el permiso `products:create` (`AUTH-002`)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "409":
          description: "Código o nombre ya en uso (`EX-005`, `EX-001`)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "422":
          description: "Membresía inexistente, moneda inexistente o inactiva, o un\
            \ origen POR ENCIMA del destino (`EX-002`, `EX-003`, `EX-006`)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /api/v1/products/{id}/deletion:
    post:
      tags:
      - Productos
      summary: Retirar un producto del catálogo
      description: |
        Retira el producto: **eliminación lógica y con motivo** (Art. V.13).
        La fila se conserva —lo vendido tiene que seguir resolviendo a lo que
        se vendió— y el producto deja de ofrecerse.

        **`POST` sobre un subrecurso y no `DELETE` con cuerpo**, igual que al
        eliminar un rol o una persona: RFC 9110 no define semántica para el
        cuerpo de un `DELETE` y un intermediario puede descartarlo, con lo que
        la petición se convertiría en un rechazo por motivo ausente que quien
        la envió no puede entender ni corregir. Y tampoco en la URL, donde el
        motivo quedaría escrito en los registros de acceso de cualquier proxy.

        **El motivo es obligatorio** y se comprueba lo primero de todo, antes
        de tocar la base.

        **El estado no se modifica al retirar.** El registro de eliminación
        conserva si el producto **estaba a la venta**: desactivarlo «de paso»
        haría que todos los registros dijeran «inactivo» y ese dato dejaría de
        significar nada.

        **Qué libera el retiro y qué no**: el **destino** del upgrade queda
        libre para que otro se active, y el **nombre** queda libre para otro
        producto. El **código no se libera nunca** — el día que una factura
        diga `UPGRADE_ORO` tiene que resolver a un solo producto para siempre.

        **No es idempotente a propósito**: retirar uno ya retirado devuelve
        `409`. Dos motivos sobre un solo hecho es evidencia contradictoria.

        No devuelve el producto: lo que se acaba de retirar no es algo que el
        sistema deba seguir ofreciendo a quien lo pidió.
      operationId: retirar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteProductRequest"
      responses:
        "204":
          description: Producto retirado.
        "400":
          description: "Identificador sin forma canónica (`VAL-001`), motivo ausente\
            \ o en blanco (`VAL-002`) o demasiado largo (`VAL-003`)"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin el permiso `products:delete` (`AUTH-002`)
        "404":
          description: No existe un producto con ese identificador (`EX-001`)
        "409":
          description: El producto ya estaba retirado (`EX-002`)
        "500":
          description: Fallo no controlado (`ERR-500`)
...
  /api/v1/commission-rates/{id}/products/{productId}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar la asociación de la tasa con un producto
      description: |
        **Es la única forma de dejar de pagar sin retirar la tasa**: sigue en el
        catálogo y disponible para otros productos.

        **El borrado es físico** y la tabla no tiene retiro lógico, de modo que
        **el registro de eliminación es lo único que queda** de que esa tasa rigió
        alguna vez sobre ese producto. De ahí el **motivo obligatorio**.

        Si esa tasa no está asociada a ese producto devuelve `404` — y no `409`,
        porque al no quedar rastro no puede distinguirse «nunca existió» de «ya se
        borró».
      operationId: desasociar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissociateProductRequest"
      responses:
        "204":
          description: Asociación retirada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: Esa tasa no está asociada a ese producto
...
components:
  schemas:
...
    DissociateProductRequest:
      type: object
      properties:
        reason:
          type: string
    DeleteCommissionRateRequest:
      type: object
      properties:
        reason:
          type: string
...
    DeleteProductRequest:
      type: object
      properties:
        reason:
          type: string
...
    ProductPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductItem"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalIsExact:
          type: boolean
        sort:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token de acceso emitido por `POST /api/v1/auth/login`. Pegue solo\
        \ el valor de `accessToken`: el prefijo `Bearer` lo añade la interfaz."
      in: header
      scheme: bearer
      bearerFormat: JWT